oauthlib_example/requirements.txt
-i https://pypi.org/simple
aiocontextvars==0.2.2 ; python_version < '3.7'
asgiref==3.2.2
asn1crypto==1.0.1
attrs==19.2.0
autobahn==19.10.1
automat==0.7.0
cffi==1.12.3
constantly==15.1.0
contextvars==2.4 ; python_version < '3.7'
cryptography==2.7
daphne==2.3.0
databases==0.2.5
dataclasses==0.6 ; python_version < '3.7'
dnspython==1.16.0
email-validator==1.0.4
fastapi==0.41.0
hyperlink==19.0.0
idna==2.8
immutables==0.10
incremental==17.5.0
oauthlib==3.1.0
passlib==1.7.1
pbr==5.4.3
pycparser==2.19
pydantic==0.32.2
pyhamcrest==1.9.0
pyjwt==1.7.1
python-multipart==0.0.5
six==1.12.0
sqlalchemy==1.3.9
starlette==0.12.9
twisted==19.7.0
txaio==18.8.1
zope.interface==4.6.0
